Split App.tsx file into multiple components under src/components
