+++ src/App.tsx
@@ -1,821 +1,16 @@
-import React, { useState, useEffect, useMemo, useCallback, useRef } from 'react';
-import { ChefHat as Chef, TrendingUp, Recycle, Brain, BarChart3, Calendar, Settings, User, Mic, IndianRupee, Lightbulb, Target, Clock, Star, Phone, BotMessageSquare, Loader2Icon, Menu, X } from 'lucide-react';
+import React, { useState, useEffect, useMemo, useCallback } from 'react';
+import { ChefHat as Chef, TrendingUp, Recycle, Brain, BarChart3, Calendar, Settings, User, Mic, IndianRupee, Lightbulb, Target, Clock, Star, Menu } from 'lucide-react';
 import { Chart as ChartJS, CategoryScale, LinearScale, PointElement, LineElement, BarElement, ArcElement, Title, Tooltip, Legend } from 'chart.js';
-import { Bar, Pie } from 'react-chartjs-2';
-import Vapi from '@vapi-ai/web';
-import { Toaster, toast } from 'sonner';
-
+import AIChef from './components/AIChef';
+import ProfitOptimizer from './components/ProfitOptimizer';
+import PerformanceAnalytics from './components/PerformanceAnalytics';
+import ZeroWaste from './components/ZeroWaste';
+import LandingPage from './components/LandingPage';
+import ChatGPT from './components/ChatGPT';
+import DailyStrategy from './components/DailyStrategy';
 
 // Register Chart.js components
 ChartJS.register(CategoryScale, LinearScale, PointElement, LineElement, BarElement, ArcElement, Title, Tooltip, Legend);
-
-// ============================================================================
-// == FEATURE COMPONENT: AI Sous-Chef (Chaat-GPT)
-// ============================================================================
-
-const Notification = ({ message, onClose }: { message: string | null, onClose: () => void }) => {
-    if (!message) return null;
-    return (
-        <div className="bg-red-100 border-l-4 border-red-500 text-red-700 p-4 rounded-lg shadow-md flex justify-between items-center animate-fade-in">
-            <span>{message}</span>
-            <button onClick={onClose} className="font-bold text-xl">&times;</button>
-        </div>
-    );
-};
-
-const LoadingSpinner = () => (
-    <div className="text-center p-4">
-        <div className="w-10 h-10 border-4 border-gray-200 border-t-orange-500 rounded-full animate-spin mx-auto mb-2"></div>
-        <p className="text-gray-600">Chaat-GPT is thinking...</p>
-    </div>
-);
-
-const RecommendationBox = ({ text }: { text: string }) => (
-    <div className="bg-blue-50 border-l-4 border-blue-500 text-blue-800 p-5 rounded-lg shadow-md animate-fade-in">
-        <h2 className="font-bold text-lg mb-2">🧠 Chaat-GPT Suggestion:</h2>
-        <p className="text-base leading-relaxed whitespace-pre-wrap">{text}</p>
-    </div>
-);
-
-const AIChef = () => {
-    const [inputText, setInputText] = useState('');
-    const [recommendation, setRecommendation] = useState('');
-    const [isLoading, setIsLoading] = useState(false);
-    const [isListening, setIsListening] = useState(false);
-    const [currentDateTime, setCurrentDateTime] = useState('');
-    const [notification, setNotification] = useState<string | null>(null);
-
-    const API_KEY = "AIzaSyASxHWMU-e4sweZohMia3iVN3vefSRh0l8"; // IMPORTANT: Replace with your key
-    const API_URL = `https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:generateContent?key=${API_KEY}`;
-
-    const recognition = useMemo(() => {
-        const SpeechRecognition = (window as any).SpeechRecognition || (window as any).webkitSpeechRecognition;
-        if (!SpeechRecognition) return null;
-        const recognitionInstance = new SpeechRecognition();
-        recognitionInstance.continuous = false;
-        recognitionInstance.lang = 'en-IN';
-        recognitionInstance.interimResults = false;
-        return recognitionInstance;
-    }, []);
-
-    useEffect(() => {
-        if (!recognition) return;
-        recognition.onresult = (event) => {
-            setInputText(event.results[0][0].transcript);
-            setIsListening(false);
-        };
-        recognition.onerror = (event) => {
-            setNotification(`Speech recognition error: ${event.error}`);
-            setIsListening(false);
-        };
-        recognition.onend = () => setIsListening(false);
-    }, [recognition]);
-
-    useEffect(() => {
-        const now = new Date();
-        const options: Intl.DateTimeFormatOptions = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
-        setCurrentDateTime(`${now.toLocaleDateString('en-US', options)} | Hyderabad, Telangana`);
-    }, []);
-
-    const getAIResponseFromGemini = async (text: string) => {
-        if (API_KEY === "YOUR_GEMINI_API_KEY") {
-            setNotification("Please add your Gemini API Key to the AIChef component.");
-            return null;
-        }
-        const prompt = `You are Chaat-GPT, an expert AI sous-chef for Indian street food vendors. Your tone is helpful, concise, and you understand Hinglish. A vendor has described their ingredients. Provide an actionable cooking adjustment. Do not start with "Okay, here's...". Get straight to the point. Vendor's input: "${text}"`;
-        const payload = { contents: [{ parts: [{ text: prompt }] }] };
-        try {
-            const response = await fetch(API_URL, {
-                method: 'POST', headers: { 'Content-Type': 'application/json' }, body: JSON.stringify(payload)
-            });
-            if (!response.ok) {
-                const errorBody = await response.json();
-                throw new Error(`API Error: ${errorBody.error.message}`);
-            }
-            const data = await response.json();
-            if (data.candidates?.[0]?.content?.parts?.[0]?.text) {
-                return data.candidates[0].content.parts[0].text;
-            } else {
-                throw new Error("No recommendation received from the AI.");
-            }
-        } catch (error) {
-            setNotification((error as Error).message);
-            return null;
-        }
-    };
-
-    const handleGetAdvice = async () => {
-        if (!inputText) {
-            setNotification('Please describe your ingredients first!');
-            return;
-        }
-        setIsLoading(true);
-        setRecommendation('');
-        setNotification(null);
-        const advice = await getAIResponseFromGemini(inputText);
-        if (advice) setRecommendation(advice);
-        setIsLoading(false);
-    };
-
-    const handleVoiceInput = () => {
-        if (!recognition) {
-            setNotification("Sorry, your browser doesn't support voice recognition.");
-            return;
-        }
-        if (isListening) recognition.stop();
-        else {
-            setIsListening(true);
-            recognition.start();
-        }
-    };
-
-    return (
-        <div className="h-full w-full flex items-start justify-center pt-8">
-            <div className="w-full max-w-2xl mx-auto space-y-6">
-                <header className="text-center">
-                    <h1 className="text-5xl font-bold text-orange-500">AI Sous-Chef</h1>
-                    <p className="text-lg text-gray-700 mt-1">Real-time Recipe Intelligence</p>
-                    <p className="text-sm text-gray-500 mt-2">{currentDateTime}</p>
-                </header>
-                <Notification message={notification} onClose={() => setNotification(null)} />
-                <div className="bg-white p-6 rounded-xl shadow-lg border border-gray-200">
-                    <h2 className="text-xl font-semibold text-gray-800">Aaj ke ingredients kaise hain?</h2>
-                    <p className="text-sm text-gray-500 mt-1 mb-4">Describe your ingredients or use the voice button.</p>
-                    <textarea id="ingredient-input" className="w-full h-28 p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-orange-400 transition"
-                        placeholder="e.g., 'Aaj ke tamatar thode khatte hain' or 'Chilies are not very spicy'"
-                        value={inputText} onChange={(e) => setInputText(e.target.value)} />
-                    <div className="flex flex-col sm:flex-row gap-3 mt-4">
-                        <button onClick={handleGetAdvice} className="w-full bg-orange-500 text-white font-bold py-3 px-4 rounded-lg hover:bg-orange-600 transition disabled:bg-orange-300" disabled={isLoading}>
-                            Get AI Advice
-                        </button>
-                        <button onClick={handleVoiceInput} className={`w-full font-bold py-3 px-4 rounded-lg transition ${isListening ? 'bg-red-500 text-white' : 'bg-gray-200 text-gray-700 hover:bg-gray-300'}`}>
-                            {isListening ? 'Listening...' : 'Record Voice 🎙️'}
-                        </button>
-                    </div>
-                </div>
-                {isLoading && <LoadingSpinner />}
-                {recommendation && !isLoading && <RecommendationBox text={recommendation} />}
-            </div>
-            <style>{`@keyframes fade-in { 0% { opacity: 0; transform: translateY(10px); } 100% { opacity: 1; transform: translateY(0); } } .animate-fade-in { animation: fade-in 0.5s ease-out forwards; }`}</style>
-        </div>
-    );
-};
-
-
-// ============================================================================
-// == FEATURE COMPONENT: Profit Optimizer
-// ============================================================================
-const ProfitOptimizer = ({
-    ingredientPrices, platesSold, menuItems,
-    handlePriceChange, handlePlatesSoldChange,
-    dailyPerformance, aiRecommendation
-}: any) => (
-    <div className="bg-white rounded-2xl shadow-sm border border-gray-200 p-4 sm:p-8">
-        <h1 className="text-3xl font-bold text-center text-green-700 mb-6">Profit Optimizer</h1>
-        <div className="bg-green-50 p-6 rounded-lg shadow-inner mb-6">
-            <h2 className="text-xl font-bold text-green-600 mb-4">Daily Ingredient Prices (₹ per kg/L)</h2>
-            <div className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-5 gap-4">
-                {Object.keys(ingredientPrices).map(ing => (
-                    <div key={ing} className="flex flex-col">
-                        <label htmlFor={ing} className="capitalize text-sm font-medium text-gray-700 mb-1">{ing}:</label>
-                        <input type="number" id={ing} name={ing} value={ingredientPrices[ing]} onChange={handlePriceChange}
-                            className="p-2 border border-green-300 rounded-md focus:ring-2 focus:ring-green-500" min="0" step="0.01" />
-                    </div>
-                ))}
-            </div>
-        </div>
-        <div className="bg-yellow-50 p-6 rounded-lg shadow-inner mb-6">
-            <h2 className="text-xl font-bold text-yellow-700 mb-4">Daily Sales Input (Plates Sold)</h2>
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-x-6 gap-y-4">
-                {menuItems.map((item: any) => (
-                    <div key={item.id} className="flex flex-col">
-                        <label className="text-sm font-medium text-gray-700 mb-1">{item.name}:</label>
-                        <div className="flex items-center space-x-2">
-                            <button onClick={() => handlePlatesSoldChange(item.id, -1)} className="p-1.5 bg-red-400 text-white rounded-full shadow-md hover:bg-red-500 transition-colors focus:outline-none focus:ring-2 focus:ring-red-300">
-                                <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" viewBox="0 0 20 20" fill="currentColor"><path fillRule="evenodd" d="M5 10a1 1 0 011-1h8a1 1 0 110 2H6a1 1 0 01-1-1z" clipRule="evenodd" /></svg>
-                            </button>
-                            <span className="flex-grow text-center text-lg font-semibold">{platesSold[item.id]}</span>
-                            <button onClick={() => handlePlatesSoldChange(item.id, 1)} className="p-1.5 bg-green-400 text-white rounded-full shadow-md hover:bg-green-500 transition-colors focus:outline-none focus:ring-2 focus:ring-green-300">
-                                <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" viewBox="0 0 20 20" fill="currentColor"><path fillRule="evenodd" d="M10 5a1 1 0 011 1v3h3a1 1 0 110 2h-3v3a1 1 0 11-2 0v-3H6a1 1 0 110-2h3V6a1 1 0 011-1z" clipRule="evenodd" /></svg>
-                            </button>
-                        </div>
-                    </div>
-                ))}
-            </div>
-        </div>
-        <div className="bg-purple-50 p-6 rounded-lg shadow-inner border-l-4 border-purple-400 mb-6">
-            <h2 className="text-xl font-bold text-purple-700 mb-2">StallOS AI Recommendation</h2>
-            <p className="text-md text-gray-800 leading-relaxed" dangerouslySetInnerHTML={{ __html: aiRecommendation.replace(/\*\*(.*?)\*\*/g, '<strong>$1</strong>') }} />
-        </div>
-        <div className="bg-gray-50 p-6 rounded-lg shadow-inner">
-            <h2 className="text-xl font-bold text-gray-700 mb-4">Today's Performance Summary</h2>
-            <div className="flex flex-col sm:flex-row justify-end items-end sm:items-center space-y-2 sm:space-y-0 sm:space-x-6 text-xl font-extrabold text-gray-800">
-                <span className="bg-gray-200 p-2 px-4 rounded-lg shadow-sm">Total Revenue: ₹{dailyPerformance.totalRevenue.toFixed(2)}</span>
-                <span className="bg-green-200 p-2 px-4 rounded-lg shadow-sm text-green-900">Total Profit: ₹{dailyPerformance.totalOverallProfit.toFixed(2)}</span>
-            </div>
-        </div>
-    </div>
-);
-
-
-// ============================================================================
-// == DYNAMIC FEATURE COMPONENT: Performance Analytics
-// ============================================================================
-const PerformanceAnalytics = ({ menuData, dailyPerformance }: { menuData: any[], dailyPerformance: { totalRevenue: number, totalOverallProfit: number } }) => {
-    
-    const profitDistributionData = useMemo(() => {
-        const positiveProfitItems = menuData.filter(item => item.itemProfit > 0);
-        return {
-            labels: positiveProfitItems.map(item => item.name),
-            datasets: [{
-                data: positiveProfitItems.map(item => item.itemProfit),
-                backgroundColor: positiveProfitItems.map(() => `hsla(${Math.random() * 360}, 70%, 60%, 0.8)`),
-                borderColor: '#ffffff',
-                borderWidth: 2,
-            }],
-        };
-    }, [menuData]);
-
-    const profitDistributionOptions: any = {
-        responsive: true, maintainAspectRatio: false,
-        plugins: {
-            legend: { position: 'right' },
-            title: { display: true, text: 'Profit Contribution by Item (Today)', font: { size: 18, weight: 'bold' } },
-            tooltip: {
-                callbacks: {
-                    label: function(context: any) {
-                        const total = context.dataset.data.reduce((sum: number, val: number) => sum + val, 0);
-                        const percentage = total > 0 ? ((context.raw / total) * 100).toFixed(1) : 0;
-                        return `${context.label}: ₹${Number(context.raw).toFixed(2)} (${percentage}%)`;
-                    }
-                }
-            }
-        },
-    };
-
-    const salesVsMarginData = useMemo(() => {
-        const sortedData = [...menuData].sort((a, b) => b.platesSold - a.platesSold);
-        return {
-            labels: sortedData.map(item => item.name),
-            datasets: [
-                { label: 'Plates Sold', data: sortedData.map(item => item.platesSold), backgroundColor: 'rgba(255, 159, 64, 0.7)', yAxisID: 'y-sales' },
-                { label: 'Profit Margin (%)', data: sortedData.map(item => item.profitMargin), backgroundColor: 'rgba(75, 192, 192, 0.7)', yAxisID: 'y-margin' },
-            ],
-        };
-    }, [menuData]);
-    
-    const salesVsMarginOptions: any = {
-        responsive: true, maintainAspectRatio: false,
-        plugins: {
-            legend: { position: 'top' },
-            title: { display: true, text: 'Sales Volume vs. Profit Margin', font: { size: 18, weight: 'bold' } },
-        },
-        scales: {
-            'y-sales': { type: 'linear', position: 'left', title: { display: true, text: 'Plates Sold' } },
-            'y-margin': { type: 'linear', position: 'right', title: { display: true, text: 'Profit Margin (%)' }, grid: { drawOnChartArea: false } },
-        },
-    };
-
-    return (
-        <div className="bg-white rounded-2xl shadow-sm border border-gray-200 p-4 sm:p-8 space-y-8">
-            <h1 className="text-4xl font-extrabold text-center text-pink-700">Live Performance Analytics</h1>
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                <div className="bg-blue-50 p-6 rounded-lg shadow-lg text-center border border-blue-200">
-                    <p className="text-lg font-semibold text-gray-700">Total Revenue (Today)</p>
-                    <p className="text-4xl font-bold text-blue-700">₹{dailyPerformance.totalRevenue.toFixed(2)}</p>
-                </div>
-                <div className="bg-green-50 p-6 rounded-lg shadow-lg text-center border border-green-200">
-                    <p className="text-lg font-semibold text-gray-700">Total Profit (Today)</p>
-                    <p className="text-4xl font-bold text-green-700">₹{dailyPerformance.totalOverallProfit.toFixed(2)}</p>
-                </div>
-            </div>
-            <div className="grid grid-cols-1 lg:grid-cols-2 gap-8">
-                <div className="bg-gray-50 p-4 rounded-lg shadow-inner h-[400px]">
-                    <Pie data={profitDistributionData} options={profitDistributionOptions} />
-                </div>
-                <div className="bg-gray-50 p-4 rounded-lg shadow-inner h-[400px]">
-                    <Bar data={salesVsMarginData} options={salesVsMarginOptions} />
-                </div>
-            </div>
-        </div>
-    );
-};
-
-// ============================================================================
-// == NEW FEATURE COMPONENT: Zero-Waste Genius
-// ============================================================================
-
-const ZeroWasteLoadingSpinner = () => (
-    <div className="flex flex-col items-center justify-center gap-4 p-8">
-      <div className="w-12 h-12 border-4 border-dashed rounded-full animate-spin border-blue-500"></div>
-      <p className="text-gray-600">Generating your genius recipe...</p>
-    </div>
-);
-  
-const ZeroWaste = () => {
-    const [leftoverIngredients, setLeftoverIngredients] = useState('');
-    const [generatedRecipe, setGeneratedRecipe] = useState<{ recipeName: string, instructions: string } | null>(null);
-    const [isLoading, setIsLoading] = useState(false);
-    const [error, setError] = useState<string | null>(null);
-    const [isListening, setIsListening] = useState(false);
-    const recognitionRef = useRef<any>(null);
-
-    useEffect(() => {
-        const SpeechRecognition = (window as any).SpeechRecognition || (window as any).webkitSpeechRecognition;
-        if (SpeechRecognition) {
-            const recognition = new SpeechRecognition();
-            recognition.continuous = false;
-            recognition.lang = 'en-IN';
-            recognition.interimResults = false;
-            recognition.onstart = () => setIsListening(true);
-            recognition.onend = () => setIsListening(false);
-            recognition.onerror = (event) => {
-                setError(`Audio error: ${event.error}. Please type ingredients.`);
-                setIsListening(false);
-            };
-            recognition.onresult = (event) => {
-                setLeftoverIngredients(event.results[0][0].transcript);
-            };
-            recognitionRef.current = recognition;
-        }
-    }, []);
-
-    const handleListen = () => {
-        if (!recognitionRef.current) return;
-        if (isListening) recognitionRef.current.stop();
-        else {
-            setLeftoverIngredients('');
-            recognitionRef.current.start();
-        }
-    };
-
-    const handleGenerateRecipe = async () => {
-        if (!leftoverIngredients.trim()) {
-            setError("Please enter your leftover ingredients.");
-            return;
-        }
-        setIsLoading(true);
-        setError(null);
-        setGeneratedRecipe(null);
-
-        const systemInstruction = `You are "Chaat-GPT", an expert culinary assistant for Indian street food vendors. You specialize in creating delicious, sellable items from leftovers to minimize waste and maximize profit. Your tone is creative and encouraging. Your task is to take a list of leftover ingredients and generate a simple, appealing recipe suitable for a street food stall. The output must be a valid JSON object with two keys: "recipeName" (a catchy, marketable name for the dish) and "instructions" (clear, step-by-step instructions for preparation, written in a simple, friendly tone, possibly using some Hinglish terms like 'tadka' or 'bhun-lo').`;
-        const prompt = `Here are my leftovers: ${leftoverIngredients}`;
-        
-        const apiKey = "AIzaSyASxHWMU-e4sweZohMia3iVN3vefSRh0l8"; // IMPORTANT: Replace with your key
-        if (apiKey === "YOUR_GEMINI_API_KEY") {
-            setError("Please add your Gemini API Key in the ZeroWaste component.");
-            setIsLoading(false);
-            return;
-        }
-        const apiUrl = `https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:generateContent?key=${apiKey}`;
-
-        try {
-            const payload = {
-                contents: [{ role: "user", parts: [{ text: systemInstruction }, {text: prompt}] }],
-                generationConfig: {
-                    responseMimeType: "application/json",
-                    responseSchema: {
-                        type: "OBJECT",
-                        properties: { "recipeName": { "type": "STRING" }, "instructions": { "type": "STRING" } },
-                        required: ["recipeName", "instructions"]
-                    }
-                }
-            };
-            const response = await fetch(apiUrl, {
-                method: 'POST', headers: { 'Content-Type': 'application/json' }, body: JSON.stringify(payload)
-            });
-            if (!response.ok) {
-                const errorBody = await response.json();
-                throw new Error(`API Error: ${errorBody.error?.message || 'Something went wrong'}`);
-            }
-            const result = await response.json();
-            if (result.candidates?.[0]?.content?.parts?.[0]?.text) {
-                setGeneratedRecipe(JSON.parse(result.candidates[0].content.parts[0].text));
-            } else {
-                throw new Error("Failed to generate a valid recipe.");
-            }
-        } catch (err) {
-            setError((err as Error).message);
-        } finally {
-            setIsLoading(false);
-        }
-    };
-
-    return (
-        <div className="h-full w-full flex items-start justify-center pt-8">
-            <div className="w-full max-w-2xl mx-auto bg-white rounded-2xl shadow-xl p-6 md:p-10 space-y-8">
-                <div className="text-center">
-                    <div className="inline-block bg-blue-100 p-3 rounded-full">
-                        <Recycle className="w-10 h-10 text-blue-500" />
-                    </div>
-                    <h1 className="text-3xl md:text-4xl font-bold text-gray-800 mt-4">Zero-Waste Genius</h1>
-                    <p className="text-gray-500 mt-2">Turn today's leftovers into tomorrow's profits!</p>
-                </div>
-                <div className="space-y-4">
-                    <div className="flex justify-between items-center">
-                        <label htmlFor="ingredients" className="block text-lg font-medium text-gray-700">What's left over today?</label>
-                        <button onClick={handleListen} disabled={isLoading}
-                            className={`p-2 rounded-full transition-colors ${isListening ? 'bg-red-500 text-white animate-pulse' : 'bg-gray-200 text-gray-600 hover:bg-gray-300'}`}
-                            title={isListening ? "Stop listening" : "Use microphone"}>
-                            <Mic className="w-6 h-6" />
-                        </button>
-                    </div>
-                    <textarea id="ingredients" rows={4}
-                        className="w-full p-4 border border-gray-300 rounded-xl focus:ring-2 focus:ring-blue-500 transition duration-200 resize-none"
-                        placeholder={isListening ? "Listening... speak your ingredients now." : "e.g., 2 boiled potatoes, half a cup of dosa batter, some mint chutney..."}
-                        value={leftoverIngredients} onChange={(e) => setLeftoverIngredients(e.target.value)} disabled={isLoading || isListening} />
-                </div>
-                <div>
-                    <button onClick={handleGenerateRecipe} disabled={isLoading || isListening}
-                        className="w-full flex items-center justify-center gap-3 bg-blue-500 text-white font-bold py-4 px-6 rounded-xl hover:bg-blue-600 focus:outline-none focus:ring-4 focus:ring-blue-300 transition-transform transform hover:scale-105 disabled:bg-gray-400">
-                        <Recycle className="w-6 h-6" />
-                        <span>{isLoading ? 'Thinking...' : 'Generate a Genius Recipe'}</span>
-                    </button>
-                </div>
-                <div className="min-h-[200px]">
-                    {isLoading && <ZeroWasteLoadingSpinner />}
-                    {error && (
-                        <div className="bg-red-100 border-l-4 border-red-500 text-red-700 p-4 rounded-lg" role="alert">
-                            <p className="font-bold">Oops!</p>
-                            <p>{error}</p>
-                        </div>
-                    )}
-                    {generatedRecipe && (
-                        <div className="bg-green-50 border-2 border-dashed border-green-400 p-6 rounded-2xl animate-fade-in">
-                            <h2 className="text-2xl font-bold text-green-800">{generatedRecipe.recipeName}</h2>
-                            <p className="text-gray-600 mt-4 whitespace-pre-wrap font-serif">{generatedRecipe.instructions}</p>
-                            <p className="text-sm text-green-600 mt-6 font-semibold">Sell this as a limited-time special to use all your stock!</p>
-                        </div>
-                    )}
-                </div>
-            </div>
-            <style>{`.animate-fade-in { animation: fade-in 0.5s ease-out forwards; } @keyframes fade-in { from { opacity: 0; transform: translateY(10px); } to { opacity: 1; transform: translateY(0); } }`}</style>
-        </div>
-    );
-};
-// ============================================================================
-// == PLACEHOLDER & LANDING COMPONENTS
-// ============================================================================
-const LandingPage = () => (
-    <div className="bg-white rounded-2xl shadow-sm border border-gray-200 p-8 h-full flex flex-col justify-center items-center">
-        <div className="text-center max-w-4xl">
-            <div className="w-20 h-20 bg-gradient-to-r from-orange-500 to-red-500 rounded-2xl mx-auto mb-6 flex items-center justify-center">
-                <Brain className="w-10 h-10 text-white" />
-            </div>
-            <h1 className="text-4xl font-extrabold text-gray-900 mb-4">
-                Welcome to StallOS
-            </h1>
-            <p className="text-xl text-gray-600 mb-8 max-w-2xl mx-auto">
-                The AI-powered operating system for street food vendors. Transform challenges into opportunities and leftovers into profit.
-            </p>
-
-            <div className="grid grid-cols-1 md:grid-cols-3 gap-6 mt-12 text-left">
-                <div className="p-6 bg-orange-50 rounded-xl border border-orange-100">
-                    <Chef className="w-8 h-8 text-orange-600 mb-3" />
-                    <h3 className="font-semibold text-gray-900 mb-2">
-                        AI Recipe Adjustments
-                    </h3>
-                    <p className="text-sm text-gray-600">
-                        Maintain consistent taste despite ingredient variations.
-                    </p>
-                </div>
-
-                <div className="p-6 bg-green-50 rounded-xl border border-green-100">
-                    <TrendingUp className="w-8 h-8 text-green-600 mb-3" />
-                    <h3 className="font-semibold text-gray-900 mb-2">
-                        Smart Profit Optimization
-                    </h3>
-                    <p className="text-sm text-gray-600">
-                        Turn market volatility into daily opportunities.
-                    </p>
-                </div>
-
-                <div className="p-6 bg-blue-50 rounded-xl border border-blue-100">
-                    <Recycle className="w-8 h-8 text-blue-600 mb-3" />
-                    <h3 className="font-semibold text-gray-900 mb-2">
-                        Zero-Waste Intelligence
-                    </h3>
-                    <p className="text-sm text-gray-600">
-                        Convert leftovers into profitable menu items.
-                    </p>
-                </div>
-            </div>
-            <p className="mt-12 text-gray-500">Select a feature from the sidebar to get started.</p>
-        </div>
-    </div>
-);
-
-// ============================================================================
-// == VAPI-POWERED VOICE COMPONENT
-// ============================================================================
-const ChatMessage = ({ role, text }: { role: 'user' | 'assistant', text: string }) => (
-  <div className={`flex items-start gap-3 my-4 ${role === 'user' ? 'justify-end' : 'justify-start'}`}>
-    {role === 'assistant' && <BotMessageSquare className="h-8 w-8 text-blue-500 flex-shrink-0" />}
-    <div className={`px-4 py-2 rounded-lg max-w-lg ${role === 'user' ? 'bg-blue-500 text-white' : 'bg-gray-200 text-gray-800'}`}>
-      <p>{text}</p>
-    </div>
-  </div>
-);
-
-const ChaatGPT = () => {
-    const [vapi, setVapi] = useState<Vapi | null>(null);
-    const [isSessionActive, setIsSessionActive] = useState(false);
-    const [isAssistantSpeaking, setIsAssistantSpeaking] = useState(false);
-    const [conversation, setConversation] = useState<{ role: 'user' | 'assistant'; text: string }[]>([]);
-    const chatContainerRef = useRef<HTMLDivElement>(null);
-
-    useEffect(() => {
-        // IMPORTANT: Replace with your Vapi Public Key
-        const vapiInstance = new Vapi('59992990-1840-4767-adf9-4fd44bda821e');
-        setVapi(vapiInstance);
-
-        vapiInstance.on('call-start', () => {
-            toast.success("Chaat-GPT is connected!");
-            setIsSessionActive(true);
-            setConversation([]);
-        });
-
-        vapiInstance.on('call-end', () => {
-            toast.info("Session ended.");
-            setIsSessionActive(false);
-            setIsAssistantSpeaking(false);
-        });
-
-        vapiInstance.on('speech-start', () => setIsAssistantSpeaking(true));
-        vapiInstance.on('speech-end', () => setIsAssistantSpeaking(false));
-
-        vapiInstance.on('message', (message) => {
-            if (message.type === 'transcript' && message.transcriptType === 'final' && message.transcript) {
-                setConversation(prev => [...prev, { role: 'user', text: message.transcript }]);
-            } else if (message.type === 'assistant-message' && message.message) {
-                setConversation(prev => [...prev, { role: 'assistant', text: message.message as string }]);
-            }
-        });
-
-        return () => { vapiInstance.stop(); };
-    }, []);
-
-    useEffect(() => {
-        if (chatContainerRef.current) {
-            chatContainerRef.current.scrollTop = chatContainerRef.current.scrollHeight;
-        }
-    }, [conversation]);
-
-    const startSession = () => {
-        if (!vapi) return;
-        toast.loading("Connecting to Chaat-GPT...");
-
-        const assistantConfig = {
-            name: "Chaat-GPT",
-            firstMessage: "Namaste! Main Chaat-GPT hoon. How can I help you manage your stall today?",
-            transcriber: { provider: "deepgram", model: "nova-2", language: "en-IN", },
-            voice: { provider: "playht", voiceId: "jennifer", },
-            model: {
-                provider: "openai",
-                model: "gpt-4-turbo",
-                messages: [{
-                    role: "system",
-                    content: `You are "Chaat-GPT," a specialized AI sous-chef and business analyst for Indian street food vendors. Your persona is helpful, concise, and friendly. You understand Hinglish (Telugu + Hindi + English) perfectly. Your goal is to help vendors maximize profit and maintain quality. You have three core functions: 1. **AI Sous-Chef:** When the user describes their ingredients (e.g., "tomatoes are sour," "chilies are mild"), provide precise, simple recipe adjustments to maintain taste consistency. Give measurements in grams, ml, and simple terms like 'chutki bhar' (a pinch). 2. **Profit Optimizer:** When the user gives you ingredient prices, calculate per-plate costs and suggest which high-margin dish to promote for the day. Be direct and give clear, actionable advice. 3. **Zero-Waste Genius:** When the user tells you their leftover ingredients, generate a simple, creative recipe to sell the next day as a "special." Name the new dish. Always be ready to switch between these roles based on the user's voice commands. Start the conversation by introducing yourself and asking how you can help.`,
-                }, ],
-            },
-        };
-
-        vapi.start(assistantConfig);
-    };
-
-    const stopSession = () => { vapi?.stop(); };
-
-    const handleButtonClick = () => {
-        if (isSessionActive) { stopSession(); } 
-        else { startSession(); }
-    };
-
-    return (
-        <div className="h-full w-full flex items-center justify-center">
-            <Toaster richColors position="top-right" />
-            <div className="w-full max-w-2xl h-[90%] lg:h-[80%] flex flex-col bg-white rounded-2xl shadow-xl border border-gray-200">
-                <header className="flex justify-between items-center p-6 border-b">
-                    <h1 className="text-2xl font-bold text-gray-800">Chaat-GPT Assistant</h1>
-                    <div className="flex items-center gap-3">
-                        <span className="text-sm font-medium text-gray-500">
-                            {isAssistantSpeaking ? "Assistant is speaking..." : "Ready"}
-                        </span>
-                        {isAssistantSpeaking && <div className="w-4 h-4 bg-blue-500 rounded-full animate-pulse"></div>}
-                    </div>
-                </header>
-
-                <div ref={chatContainerRef} className="flex-1 p-6 overflow-y-auto">
-                    {conversation.length === 0 ? (
-                        <div className="flex flex-col items-center justify-center h-full text-gray-500">
-                            <BotMessageSquare className="h-16 w-16 mb-4" />
-                            <p>Click "Start Session" to talk to Chaat-GPT.</p>
-                            <p className="text-sm mt-2">e.g., "Aaj ke tamatar thode khatte hain."</p>
-                        </div>
-                    ) : (
-                        conversation.map((msg, index) => (
-                            <ChatMessage key={index} role={msg.role} text={msg.text} />
-                        ))
-                    )}
-                </div>
-
-                <div className="flex justify-center items-center p-6 border-t">
-                    <button
-                        onClick={handleButtonClick}
-                        disabled={!vapi}
-                        className={`flex items-center gap-3 px-8 py-4 text-lg font-semibold text-white rounded-full transition-all duration-300 shadow-lg ${!vapi ? 'bg-gray-400 cursor-not-allowed' : ''} ${isSessionActive ? 'bg-red-500 hover:bg-red-600' : 'bg-blue-500 hover:bg-blue-600'}`}>
-                        {isSessionActive ? (
-                            <><Phone className="h-6 w-6" /> End Session</>
-                        ) : (
-                            !vapi ? (
-                                <><Loader2Icon className="h-6 w-6 animate-spin" /> Initializing...</>
-                            ) : (
-                                <><Mic className="h-6 w-6" /> Start Session</>
-                            )
-                        )}
-                    </button>
-                </div>
-            </div>
-        </div>
-    );
-};
-
-
-// ============================================================================
-// == NEW FEATURE: DAILY SALES STRATEGY
-// ============================================================================
-const generateUUID = () => crypto.randomUUID();
-
-const formatRecommendation = (text: string | null) => {
-    if (!text) return null;
-    const lines = text.split('\n').filter(line => line.trim() !== '');
-    const finalElements = [];
-    let currentListItems: string[] = [];
-    let listKeyCounter = 0;
-
-    lines.forEach((line) => {
-        const trimmedLine = line.trim();
-        const isListItem = trimmedLine.startsWith('* ') || trimmedLine.startsWith('- ') || /^\d+\.\s/.test(trimmedLine);
-
-        if (isListItem) {
-            const content = trimmedLine.replace(/^(\* |\- |\d+\.\s)/, '').trim();
-            currentListItems.push(content);
-        } else {
-            if (currentListItems.length > 0) {
-                finalElements.push(
-                    <ul key={`ul-${listKeyCounter++}`} className="list-disc list-inside space-y-2 text-gray-700">
-                        {currentListItems.map((item, idx) => (
-                            <li key={`li-${listKeyCounter}-${idx}`}>{item}</li>
-                        ))}
-                    </ul>
-                );
-                currentListItems = [];
-            }
-            finalElements.push(<p key={`p-${finalElements.length}`} className="mt-2 text-gray-700 leading-relaxed">{trimmedLine}</p>);
-        }
-    });
-
-    if (currentListItems.length > 0) {
-        finalElements.push(
-            <ul key={`ul-${listKeyCounter++}`} className="list-disc list-inside space-y-2 text-gray-700">
-                {currentListItems.map((item, idx) => (
-                    <li key={`li-${listKeyCounter}-${idx}`}>{item}</li>
-                ))}
-            </ul>
-        );
-    }
-    return <>{finalElements}</>;
-};
-
-const DailyStrategy = () => {
-    const [userId] = useState(generateUUID());
-    const [dailyItemsData, setDailyItemsData] = useState([{ id: generateUUID(), tiffinItem: '', ingredientPrice: '', dailySales: '', perPlateProfit: '' }]);
-    const [aiRecommendation, setAiRecommendation] = useState('Enter data for your tiffin items and click "Get Strategy" to receive a daily sales strategy.');
-    const [isGenerating, setIsGenerating] = useState(false);
-    const [message, setMessage] = useState('');
-
-    const tiffinItems = ['Idli', 'Dosa', 'Vada', 'Upma', 'Poori', 'Pongal', 'Uthappam', 'Kesari'];
-
-    const addTiffinItem = () => {
-        setDailyItemsData([...dailyItemsData, { id: generateUUID(), tiffinItem: '', ingredientPrice: '', dailySales: '', perPlateProfit: '' }]);
-    };
-
-    const removeTiffinItem = (id: string) => {
-        setDailyItemsData(dailyItemsData.filter(item => item.id !== id));
-    };
-
-    const handleItemChange = (id: string, field: string, value: string) => {
-        setDailyItemsData(dailyItemsData.map(item => item.id === id ? { ...item, [field]: value } : item));
-    };
-
-    const getAiSuggestion = useCallback(async () => {
-        const isValid = dailyItemsData.every(item => item.tiffinItem && item.tiffinItem !== 'Select an item' && item.ingredientPrice !== '' && item.dailySales !== '' && item.perPlateProfit !== '');
-
-        if (dailyItemsData.length === 0) {
-            setMessage('Please add at least one tiffin item to get a strategy.'); return;
-        }
-        if (!isValid) {
-            setMessage('Please fill in all input fields and select a Tiffin Item for all added entries.'); return;
-        }
-
-        setIsGenerating(true);
-        setMessage('');
-        setAiRecommendation('Generating strategy...');
-
-        try {
-            let itemDetails = dailyItemsData.map(item => `- Tiffin Item: ${item.tiffinItem}\n  - Daily Ingredient Price: ${item.ingredientPrice} INR\n  - Daily Sales (Plates): ${item.dailySales}\n  - Per Plate Profit: ${item.perPlateProfit} INR\n`).join('');
-            const prompt = `As a sales strategist for a Tiffin Hotel, analyze the following daily data for various tiffin items:\n${itemDetails}\nBased on this information, provide a specific, actionable daily sales strategy. Consider pricing adjustments, marketing focus (e.g., promotions, highlighting specific items), menu recommendations, and operational efficiency improvements. Present the strategy in a clear, concise bullet-point format, with each point being a distinct actionable step. Ensure each point starts on a new line.`;
-            const payload = { contents: [{ role: "user", parts: [{ text: prompt }] }] };
-            const apiKey = "AIzaSyASxHWMU-e4sweZohMia3iVN3vefSRh0l8"; // IMPORTANT: Replace with your key
-            const apiUrl = `https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:generateContent?key=${apiKey}`;
-
-            const response = await fetch(apiUrl, {
-                method: 'POST', headers: { 'Content-Type': 'application/json' }, body: JSON.stringify(payload)
-            });
-            
-            if (!response.ok) {
-                const errorBody = await response.json();
-                throw new Error(`API Error: ${errorBody.error?.message || 'Something went wrong'}`);
-            }
-            const result = await response.json();
-            if (result.candidates?.[0]?.content?.parts?.[0]?.text) {
-                setAiRecommendation(result.candidates[0].content.parts[0].text);
-            } else {
-                setAiRecommendation('Failed to get a strategy. Please try again.');
-            }
-        } catch (err: any) {
-            setAiRecommendation(`An error occurred: ${err.message}`);
-            setMessage('An error occurred. Please check your network connection and API key.');
-        } finally {
-            setIsGenerating(false);
-        }
-    }, [dailyItemsData]);
-
-    return (
-        <div className="bg-gradient-to-br from-indigo-50 to-purple-50 p-4 md:p-6 font-sans">
-            <div className="w-full max-w-4xl mx-auto bg-white shadow-xl rounded-2xl p-6 sm:p-8 space-y-8">
-                <h1 className="text-4xl font-extrabold text-center text-gray-900">Daily Sales Strategy</h1>
-                <div className="text-center text-sm text-gray-600 p-2 bg-indigo-50 rounded-lg">
-                    <p>Session ID: <span className="font-mono text-indigo-800 break-all">{userId}</span></p>
-                    <p className="text-xs text-gray-500 mt-1">Note: This is a temporary ID. Data is not saved after refresh.</p>
-                </div>
-                <div className="space-y-6">
-                    {dailyItemsData.map((item, index) => (
-                        <div key={item.id} className="border border-indigo-200 rounded-lg p-4 shadow-sm bg-indigo-50">
-                            <div className="flex justify-between items-center mb-4">
-                                <h3 className="text-lg font-semibold text-gray-800">Tiffin Item #{index + 1}</h3>
-                                {dailyItemsData.length > 1 && (
-                                    <button onClick={() => removeTiffinItem(item.id)} className="px-3 py-1 bg-red-500 text-white rounded-md hover:bg-red-600 transition duration-150 ease-in-out">
-                                        Remove
-                                    </button>
-                                )}
-                            </div>
-                            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4">
-                                <div>
-                                    <label htmlFor={`tiffinItem-${item.id}`} className="block text-sm font-medium text-gray-700 mb-1">Tiffin Item</label>
-                                    <select id={`tiffinItem-${item.id}`} value={item.tiffinItem} onChange={(e) => handleItemChange(item.id, 'tiffinItem', e.target.value)} className="w-full px-4 py-2 border border-gray-300 rounded-lg shadow-sm focus:ring-indigo-500 focus:border-indigo-500 transition duration-150 ease-in-out">
-                                        <option value="">Select an item</option>
-                                        {tiffinItems.map((tiffin) => (<option key={tiffin} value={tiffin}>{tiffin}</option>))}
-                                    </select>
-                                </div>
-                                <div>
-                                    <label htmlFor={`ingredientPrice-${item.id}`} className="block text-sm font-medium text-gray-700 mb-1">Ingredient Price (INR)</label>
-                                    <input type="number" id={`ingredientPrice-${item.id}`} value={item.ingredientPrice} onChange={(e) => handleItemChange(item.id, 'ingredientPrice', e.target.value)} placeholder="e.g., 5000" className="w-full px-4 py-2 border border-gray-300 rounded-lg shadow-sm focus:ring-indigo-500 focus:border-indigo-500" />
-                                </div>
-                                <div>
-                                    <label htmlFor={`dailySales-${item.id}`} className="block text-sm font-medium text-gray-700 mb-1">Daily Sales (Plates)</label>
-                                    <input type="number" id={`dailySales-${item.id}`} value={item.dailySales} onChange={(e) => handleItemChange(item.id, 'dailySales', e.target.value)} placeholder="e.g., 150" className="w-full px-4 py-2 border border-gray-300 rounded-lg shadow-sm focus:ring-indigo-500 focus:border-indigo-500" />
-                                </div>
-                                <div>
-                                    <label htmlFor={`perPlateProfit-${item.id}`} className="block text-sm font-medium text-gray-700 mb-1">Per Plate Profit (INR)</label>
-                                    <input type="number" id={`perPlateProfit-${item.id}`} value={item.perPlateProfit} onChange={(e) => handleItemChange(item.id, 'perPlateProfit', e.target.value)} placeholder="e.g., 25" className="w-full px-4 py-2 border border-gray-300 rounded-lg shadow-sm focus:ring-indigo-500 focus:border-indigo-500" />
-                                </div>
-                            </div>
-                        </div>
-                    ))}
-                    <div className="flex justify-center">
-                        <button onClick={addTiffinItem} className="px-6 py-2 bg-green-600 text-white rounded-xl font-semibold hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2 transition">
-                            Add Another Tiffin Item
-                        </button>
-                    </div>
-                </div>
-
-                {message && (<div className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded-lg relative" role="alert"><span className="block sm:inline">{message}</span></div>)}
-
-                <div className="flex justify-center">
-                    <button onClick={getAiSuggestion} disabled={isGenerating || dailyItemsData.length === 0}
-                        className={`px-8 py-3 rounded-xl font-semibold text-white shadow-lg transition duration-300 ease-in-out ${isGenerating || dailyItemsData.length === 0 ? 'bg-gray-400 cursor-not-allowed' : 'bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 transform hover:scale-105'}`}>
-                        {isGenerating ? 'Generating...' : 'Get Daily Strategy'}
-                    </button>
-                </div>
-
-                <div className="bg-indigo-50 p-6 rounded-xl shadow-inner border border-indigo-200 min-h-[150px]">
-                    <h2 className="text-2xl font-bold text-gray-800 mb-4">AI Sales Strategy</h2>
-                    {isGenerating ? <LoadingSpinner/> : formatRecommendation(aiRecommendation)}
-                </div>
-            </div>
-        </div>
-    );
-};
 
 // ============================================================================
 // == MAIN APP COMPONENT (State Manager and Router)
@@ -928,7 +123,7 @@ function App() {
             case 'analytics': 
                 return <PerformanceAnalytics menuData={menuData} dailyPerformance={dailyPerformance} />;
             case 'zero-waste': return <ZeroWaste />;
-            case 'chaat-gpt': return <ChaatGPT />;
+            case 'chaat-gpt': return <ChatGPT />;
             case 'daily-strategy': return <DailyStrategy />;
             default: return <LandingPage />;
         }
